% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsendinblue-package.R
\docType{package}
\name{rsendinblue-package}
\alias{rsendinblue}
\alias{rsendinblue-package}
\title{rsendinblue: R Client for the SENDINBLUE API}
\description{
Provides an R client for the sendinblue API 
    <https://developers.sendinblue.com>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rsendinblue}
  \item Report bugs at \url{https://github.com/FRBCesab/rsendinblue/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
