% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsendinblue-package.R
\docType{package}
\name{rsendinblue-package}
\alias{rsendinblue}
\alias{rsendinblue-package}
\title{rsendinblue: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several 
    lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rsendinblue}
  \item Report bugs at \url{https://github.com/FRBCesab/rsendinblue/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
